update kiểu dữ liệu của id là Long

+++ premier_league/src/main/java/com/example/premier_league/repository/ITeamRepository.java
@@ -8,6 +8,5 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ITeamRepository extends JpaRepository<Team, Integer> {
-
+public interface ITeamRepository extends JpaRepository<Team, Long> {
 }
